Removes koa-mount, adds middleware that provides the same functionality
We where not really using the main functionality of it, to be able to
run multiple Koa application on the same port. The added middleware
gives us the feature that we want, being able to run the application on
a specified basepath.

This change is done primarily to make it easier to interact with the Koa
application, for instance when using koa-session.